Fixed #801: Unable to share videos

+++ lib/class/catalog.class.php
@@ -1900,6 +1900,9 @@ abstract class Catalog extends database_object
 
         $catalog = self::create_from_id($catalog_id);
 
+        if (!$catalog->id)
+            return false;
+
         $sql = 'DELETE FROM `catalog_' . $catalog->get_type() . '` WHERE catalog_id = ?';
         $db_results = Dba::write($sql, array($catalog_id));
